+++ .gitignore
@@ -14,4 +14,4 @@ artifacts
 tmp
 package-lock.json
 /*.html
-/build
+/js/resolc.*
